pyproject.toml
[project]
name = "risk_atlas_nexus"
version = "0.0.3"
authors = [
    { name = "Elizabeth Daly", email = "elizabeth.daly@ie.ibm.com" },
    { name = "Dhaval Salwala", email = "dhaval.vinodbhai.salwala@ibm.com" },
    { name = "Frank Bagehorn", email = "fba@zurich.ibm.com" },
    { name = "Luis Garces-Erice", email = "lga@zurich.ibm.com" },
    { name = "Sean Rooney", email = "sro@zurich.ibm.com" },
    { name = "Inge Vejsbjerg", email = "ingevejs@ie.ibm.com" },
]
description = "Risk Atlas Nexus provides tooling to help bring together disparate resources related to governance of foundation models. "
readme = "README.md"
requires-python = ">=3.11, <3.12"
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: AI Risks and governance",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "cymple",
    "linkml",
    "linkml_runtime",
    "pydantic",
    "requests",
    "rich",
    "sssom",
    "txtai",
    "tqdm",
    "logzero",
    "python-dotenv",
    "datasets",
    "openai>=1.0",
    "txtai",
    "isort",
    "pre-commit"
]


[project.optional-dependencies]
wml = ["ibm-watsonx-ai"]
ollama = ["ollama"]
vllm = ["vllm", "xgrammar"]
docs = [
    "mkdocs-material",
    "mkdocs-jupyter",
    "mkdocs-click",
    "mkdocstrings",
    "griffe_inherited_docstrings",
    "griffe-pydantic",
    "mkdocs-awesome-nav",
]

[tool.isort]
skip_glob = ["vrisk-atlas-nexus/*"]
profile = "black"
line_length = 88
remove_redundant_aliases = true
group_by_package = true
combine_star = true
lines_after_imports = 2
